+++ src/Payline/Recurring.class.php
@@ -1,0 +1,28 @@
+<?php
+/*
+ * This file is part of the Payline package.
+ *
+ * (c) Monext <http://www.monext.net>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace Payline;
+
+class Recurring
+{
+
+    public $firstAmount;
+
+    public $amount;
+
+    public $billingCycle;
+
+    public $billingLeft;
+
+    public $billingDay;
+
+    public $startDate;
+
+    public $endDate;
+}