+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+* 4.45.1 (2016-03-10)
+
+  * add of Recurring class
+  * require any 1.* monolog/monolog version
+  * date_default_timezone_set("Europe/Paris") in PaylineSDK class constructor
+  * add widget related constants (js and css url path)
+
 * 4.45 (2016-01-22)
 
   * add of softDescriptor payment attribute (Payment class)